made todo a component
